kraken: modify the way we get impacted stops in a section
Currently we are getting the smallest set of stop_points impacted by a
section. This commit change the way we do it.
Now we :
  - get a set of ranks impacted. This allow us to impacted a stop_time
  only if its stop_point really is in a line_section (lollipop vj corner
  case)
  - handle the case where consecutive stop_points in the same area are
  at the start / the end of the section. We get the largest set of
  stop_points in these areas.

This lead to a few changes to both the way we impact vj to work with
ranks (must get the base stop_time to each stop_time to compare its
original rank) and how we register the impact on stop_points (must loop
on the base_vj to get the stop_points of each rank since they are not
stored in ImpactedVJ anymore).