kraken: fix tests for ranks
Since we change the way we impact line_section two tests needed changes.
The first one was testing the method returning the impacted_stops. Since
it returns rank now change the test and add a few more cases.
One of the apply_disruption test was failing too because we now properly
impact the shortest_section. Since it's two loops both the first
iteration of the first loop and the last iteration of the second loop
are not impacted anymore (before there were because when a stop_point
was impacted all its occurences were impacted).