ed: do not import stops from pathways
For now we do not want to import station entrance, generic nodes and
boarding position as we can't do anything with them currently.
This PR refacto the StopHandler from the gtfs parser and fusio parser to
ignore them.
FusioParser had to evolve since the `location_type` aren't matching
between gtfs and ntfs