Add admins fetched from Cities db to Rtree for significant speed up.
Requesting Cities db is significantly slow (up to 200ms). We can't
affort to do this for all POI that can possibly be loaded.

Adding the Admins comming from Citeis to the Rtree used for (GeoRef admins
lookup) significantly improve performances.

In the case of Furet coverage, ed2nav was taking up to 6h. With this
improvement, it takes ~15 minutes.