Fix prolongation trip that goes back in time
Because of allowing pick-up and drop-off at the extremity of a vehicle journeys with a
prolongation, we were able to go back in time.

This PR allows pick-up at the first
stop time and drop-off at the last stop time only for vehicle journeys
that has a prolongation with another one at a different stop point.

This solves https://github.com/CanalTP/artemis/pull/378