Change Labels to work on a per Journey Pattern Point base
Because `Labels` are currently based on a single stop point, when we try
to find the best time for a vehicle journey at this stop point, we might
"forget" another vehicle journey on a different route that arrive later.

By working on a JPP base, we keep in memory the earliest arrial for all
the different routes at the same stop point.