Count the number of RT messages in queue to be retrieved
Consume messages until this number (or max) is reached.
This allows for larger timeouts, as Kraken will stop at that number, and
not continue as long as messages arrive (which is constantly the case for
some contributors).

Also:
* Common queue declaration (to avoid bad-redeclaration)
* Check (only for log currently) that consumer is unique
* minor cleanups of useless variables

Note:
DeclareQueue() uses DeclareQueueWithCounts() under the cover in
SimpleAmpqClient, which allows for the common queue declaration refactor.

JIRA: https://navitia.atlassian.net/browse/NAV-1879